feat: better code style
