+++ lerna.json
@@ -1,6 +1,6 @@
 {
   "lerna": "2.11.0",
   "packages": ["packages/*"],
-  "npmClient": "yarn",
+  "npmClient": "cnpm",
   "version": "0.0.0"
 }
